refactor: replace threads reply forms with <tuleap-pullrequest-new-comment-form/>
part of story #32316 Update comments

Context:

A comment has a reply form which can be toggled when the user clicks on
the [Reply] button. This form was generated by a template function and
managed by the component's controller. The writing zone was the
component's one, and served only this purpose.

In this user story, we are implementing an edition feature, which
requires to use the writing zone as well. Sadly, it is not possible as
it is to use the same writing zone for two purposes.

The refactoring in this contribution aims to replace the good old
template function by the already existing new-comment-form component.
This component is used to create global and inline comments, both in the
overview and changes pane. It looks and behaves exactly like the current
reply form, except it does not create replies to a particular comment.

Using "creation contexts" makes it possible to create new comments and
replies, e.g:
- we want to create a new global comment
- we want to create a new inline comment on a particular file line
- we want to reply to a global comment
- we want to reply to an inline comment

These actions are made by two different object: one to save new
comments, one to save replies. And that's it.

Thanks to this refactoring, we'll be able to have a dedicated writing
zone to edit the comments.

How to test:
- You can add global and inline comments
- You can answer to global and inline comments
- Inline comments are created at the right place in the files
- Replies are created using the current user

Change-Id: I8fad99b5920dbf4271a8bfa9eb8982ef82f28a6f