refactor: extract action events from <tuleap-pullrequest-comment/>
part of story #32316 Update comments

Context:

In the begining of the work on pull-requests comments, action events and
regular comments were merged together. It was kind of fine at this time
because we only needed to display them.

With the addition of features to the comments, such as the
ability to reply to a comment to create a thread, or edit an already
existing comment, it does not make sense anymore.

The goal of this refactoring is to extract what's related to action
events out of the <tuleap-pullrequest-comment/> component, in a
dedicated custom element. The benefits are:
- It is easier to build a PullRequestCommentPresenter.
- There is no need to check the type of item we are displaying anymore
  (except for global vs inline comments).
- We can handle action events messages and their translations inside the
  dedicated custom element.
- Action events presenters now contain only a few attributes instead of a
  ton of useless keys containing default values.

Nevertheless, the rendering of the timeline is a bit more complex given
that we now have to differentiate action events and comments, display them
in the right order and instantiate the right custom element.

How to test:
--> No functional change expected

Change-Id: I67cddc46f0ee84ca39a169b295f05b5e687b1759