feat: add a nullable last_edition_date column to comments tables
part of story #32316 Update comments

This column is needed to store the last edition date of global and
inline comments. When a comment has never been edited, this attribute is
null, so we don't have to handle date comparisons frontend-side to know
if we display it or not.

How to test:
- run forgeupgrade
--> Great success
--> Both tables have a last_edition_date column with a null value for
    each n-uplets

Change-Id: I07a4702b442caff5b08a7dc96417b7be2035c612