fix: passwordless login form leak information about users
Part of story #33005 login with only my passkey

Login form leak information about who has registered a passkey or not.
To prevent that, when a users has no passkey, form send a false key then
check of authentication will failed.

*Testing:*

Try to authenticate with a user with no passkey, it will ask to use one,
send result to server and failed authentication.
For users with registered passkey nothing change.

Change-Id: I3a3df2028af44de23aae88d58a3194ff7858412f