chore: Move stubs in the correct folders
No functional change. CI should be happy.

Notes: It's better to have "Tests" in the namespace of stubs, as they
are meant to be used only for unit tests and should never be used for
production code. Hence, the rename of the Git stub namespace.
Some stubs in Core also were in the wrong folder, which also made them
skip Psalm validation.

part of story #32316 Update comments

Change-Id: I1f09c378953411ed4ad7b6621dc0f36147803aa5