feat: add feature flag to block passwordless login form
Part of story #33005 login with only my passkey

At the current state of WebAuthn, it's not possible to have a
passwordless login without leaking some informations about users.

To avoid that, we keep the login form, but behind a hidden feature flag.

*Testing:*

You can change the value of feature flag via tuleap comand line:

  tuleap config-set feature_flag_enable_passwordless_login {1|0}
  tuleap -c

If set to 1 -> passwordless login is enabled
If set to 0 -> passwordless login is disabled

Change-Id: I165da8534ba673b1fa268b17607c0f668ac6bdcb