feat: Remove tools from sidebar when collapsed
When sidebar is collapsed, now we don't display anything but platform
logo, Tuleap logo, and expand button.

This is because we don't know yet how to display complex upcoming data
(submilestones, grouped tools, …) in a collapsed sidebar. As a shortcut,
we display the bare minimum.

Implementation note: Even if SidebarHeader is not displayed when sidebar
is collapsed `v-if="!is_sidebar_collapsed"`, we still pass status of the
collapse as a prop `v-bind:is_sidebar_collapsed="is_sidebar_collapsed"`
because we don't know if current proposal is the right move yet (like
for the removal of the collapse). So we keep for now the logic with
collapse inside SidebarHeader and its subcomponents until we have proper
feedbacks.

Part of story #35105: collapse sidebar again

Change-Id: Ic6da5b5c959209bc6a1d0d0c173723bec9fb643f