feat: introduce calendar event config
Future calendar events will be send alongside artifact notification only
if the underlying tracker is configured to send those events.

This commit introduce this brand new configuration. We like baby steps,
so this contribution is not really visual. You should have a new table
`plugin_tracker_calendar_event_config` and if you insert data for a
tracker like `(123, 1)` (to allow sending of events for tracker with
id=123), and you create another tracker based on it, then the new
tracker will inherit the config (you should have `(124, 1)` if 124 is
the id of the new tracker).

No functional changes elsewhere.

Part of story #35094: have "calendar" events attached to tracker email notification

Change-Id: Ic26828b6d6a7badd8cb31e6639385f0c2fd3858b