feat: Check presence of semantic title during notif
If tracker is not configured to send calendar events, then nothing
change.

If tracker is configured to send calendar events and has title semantic
configured, then in the logs /var/log/tuleap/mail_syslog when the update
of an artifact induces a mail notification, we have:

> Tracker is configured to send calendar events alongside notification
> No calendar event for this changeset

(Actual event calendar will come in a dedicated contribution)

If tracker is configured to send calendar events but does not have title semantic
configured, then in the logs /var/log/tuleap/mail_syslog when the update
of an artifact induces a mail notification, we have:

> Tracker is configured to send calendar events alongside notification
> The tracker does not have title semantic, we cannot build calendar events

Implementation note: Not cluttering EmailNotificationTask forces us to
declare an unused (yet) MailAttachment. It will be used later when we
have minimal background checks before sending the event as attachment.

To configure tracker you have to manually update
plugin_tracker_calendar_event_config table.

Part of story #35094: have "calendar" events attached to tracker email notification

Change-Id: I8f092a9df157e47c320b289fd80a6a63b1a524e5