feat: Check permissions for title semantic
To build the calendar event sent in notification, we need a summary. We
take the title semantic for that. However if user who will receive the
notification is not able to read the underlying field, then we should
not build the calendar event. Appropriate logs are added in that case
with information about the user (id and email, useful for mailing lists)
in order to troubleshoot the issues.

The check of permission is not performed in case of global notification
with check perms = off.

Implementation note: TestLogger::hasDebugThatContains is handy, but
true/false does not help to know which debug information is missing (and
why) when we have many checks and failing tests. Therefore we now check
the whole content of records.

Part of story #35094: have "calendar" events attached to tracker email notification

Change-Id: I1d6ed93c7047a2181f56a5cfc0067c25d4453718