feat: XML Import/Export calendar event config
When a tracker is configured to send calendar event config (reminder:
only via DB for now), then when it is exported (tracker » admin » xml
export), in the generated XML file there is an attribute to specify this
config.

When the XML file is imported back in trackers » create new tracker,
then the new tracker is configured to send calendar event as well. (you
can assert that by exporting the XML structure of the tracker, or by
looking at the DB).

If you update the XML file and replace the value attribute to 0
(`should_send_event_in_notification="0"`) and you import it to create a
new tracker, then this new tracker is not configured to send calendar
event.

Implementation note: first try to implement this feature was to plug
ourselves in `TrackerCreationSettings`, however this one is used in
`\TrackerFactory::create()` which already handle the notification
settings via `NotificationSettingsDuplicator`. By continuing this path,
we would have end up updating twice the config: one via
`NotificationSettingsDuplicator` and one via `TrackerCreationSettings`.
Since it is the first time that we deal with notification settings in
tracker XML import/export (except legacy ones) it seems natural to have
a dedicated object/process `TrackerCreationNotificationSettings` to deal
with those attributes during XML import.

Part of story #35094: have "calendar" events attached to tracker email notification

Change-Id: I3894decc97492cd4636e993b79dae7e783a108c8