fix: project from source repository should be used when extracting reference
When user creates a new comment and reference an artifact into it, then
the reference is extracted from the `source` repository. However, when
the user updates his comment, the extracted reference is from the
`destination` repository.
The issue is we extract the reference from potentially two different
project in the case of a pull request from a fork.

To test:
 - Enable the feature flag: `tuleap config-set feature_flag_allow_pull_requests_comments_edition 1`
 - Create a fork of your git repository into diffrent project
 - Create a new pull request in the fork with a destination branch which is from the
   other project
 - Create a new comment with an artifact refrence
 - Edit the comment with a reference  with the API explorer
=> The link of the artifact refrenced should be in the source project
(the fork)

Part of story #32316 Update comments

Change-Id: Ia454c3903ff7483057cbb9b5244dd5c34ce6d7ba