feat: Send calendar event alongside notif
When every criteria are met, an event is sent in attachment of an
artifact notification.

Note for testers: if your platform uses mailhog to catch emails, this
tool does not display attachments (you can assert that event is in the
source of the email, and ensure its content if you can read base64).
You may want to disable mailhog in `/etc/postfix/main.cf` to receive the
mail in a real MUA:
```diff
- relayhost = mailhog:1025
+ #relayhost = mailhog:1025
```
(don't forget to restart postfix, and make sure that your container can
access the Internet)

Part of story #35094: have "calendar" events attached to tracker email notification

Change-Id: If1b8a3878ebd6dd99407d6cad3a76def3fa6e095