chore(el9): refactor generation of SVN default block
In order to have some control on the way the default svn block is
generated (to remove default [/]), let's start by extracting the code
from BackendSVN so we can re-use it elsewhere eventually.

As a side-effect, the ignominious LDAP_BackendSVN is gone.

Psalm baseline is updated because of usage of deprecated dao in the LDAP
part and I thought it was not part of this war to rewrite those DAO as
well.

It's a pure refactoring, nothing should change. You can test it by
adding/removing people from projects & groups and ensure the
.SVNAccessFile is still consistent.

Part of: request #31541 svn 1.10+ no longer supports duplicate sections

Change-Id: Ie01a654deefe015b3985a7e0e47724bbfde0e9a5