test: adjust tests since feature is now implemented
One test was saying that the feature was not implemented yet which is
not anymore the case.

Took the opportunity to add more tests on the content of the attachment.
This is to:
* ensure that if we change the underlying lib, we still have important
  information
* document why we need METHOD:REQUEST

No functional changes, tests should pass.

Part of story #35094: have "calendar" events attached to tracker email notification

Change-Id: I999d20a58515bad480731deebd1aaf174ebcc53c