story #35103: upgrade to MediaWiki 1.39
Extensions and skin have been bumped to compatibles version.
The Math extension is now integrated in the MediaWiki base tarball.

Since it is not possible to upgrade MediaWiki 1.23 to MediaWiki 1.39
directly we need to keep a MediaWiki 1.35 instance for the time being.

In order to test, build the packages with
`nix-build ./plugins/mediawiki_standalone/additional-packages/tuleap-mediawiki.nix`
and copy them inside the container. Remove the old package with `rpm -e --nodeps mediawiki-tuleap-flavor`
and install the new ones with `rpm -Uvh --nodeps ...`.

Change-Id: I1a540ac66c053dda0c20e1327a51cb51a1d8de25