fix: remove METHOD:REQUEST
While investigating on the POC for calendar events, we used the
METHOD:REQUEST in order to have convenient Yes/Maybe/No CTA in MUAs
(gmail at least).

However, since we are planning to add `noreply@…` as organizer of the
event, every time a user click on Yes/Maybe/No of the event, an email
will be sent to this address. Depending on the configuration of the
address (that we cannot control, it is up to platform administrators),
the mail can end up in error and user will receive a message about
delivery failure.

In order to avoid that, we don't anymore use METHOD:REQUEST. Gmail
displays a link to add the event to the calendar, this is less user
friendly than Yes/Maybe/No but there is less chance of bad situations.

Part of story #35094: have "calendar" events attached to tracker email notification

Change-Id: Ie9b6109dc7dfee48cc7d02ba64d5746dd7ec453c