chore: fix translation issues in main tracker po
Poedit offers comprehensive safety checks: if source strings start with
a capital letter, there is a high chance that the translation should
also start with a capital letter. The same for punctuation mark, it
should be the same in source and translation.

Contributing to those files that contains a lot of errors is painfull
because new translations for the current contribution are lost behind
previous translation issues.

This contribution fixes those issues to make life easier for the next
one.

Miffed during story #35094: have "calendar" events attached to tracker email notification

Change-Id: Iac304b6ae59913c2fdd397500ad066ccbc6620a1