feat: Display time in duration end date
Part of story #35094 have "calendar" events attached to tracker email notification

In a tracker with a timeframe based on start date and duration, when you
create an artifact with start date 25/11 and duration 2, Tuleap will
show an end date to 27/11. But when you receive the calendar event, it
last until 26/11.

In fact Tuleap end date is 27/11 at midnight, so the calendar event is
right. To remove this ambiguity we now display time next to duration
instead of just the date.

Change-Id: Iad81aed8961d6a47a72d53276f56d9457748cf83