feat: Introduce milestones in sidebar config
This introduces the config table in db in order to be able to start
working on milestones in sidebar. Since feature is not ready yet, it is
not possible to update this settings via the UI, you have to manually
update the database.

Though, if milestones in sidebar are allowed, then we can see the
setting (readonly) in project » backlog » administration. For now a
simple checkbox is displayed, it will be adjusted later after Design
Team feedbacks.

The setting is guarded by a feature flag, no matter the config in db, if
feature flag is off (default is on), then milestones will not appear in
sidebar.

To set/unset the feature flag:
tuleap config-set feature_flag_allow_milestones_in_sidebar 0

Part of story #35104: Milestones below Backlog

Change-Id: I1a25765b92bc4543dc3d2dab435d00c6c19d829c