Closes request #35156: Drop old unix_pw column in the user table
Follow up to 97d0bbeb5d5e382ae7e9635741031a5662d680fd.

The `unix_pw` contains sensitive information, we should not keep it and
we are now pretty sure we will not need to rollback the change.
Took the opportunity to also remove columns related to the change in the
user table.

No functional changes.

Change-Id: I59acdad1b16add5947861ba13a50330df5ca4cd0