feat: XML import/export of milestones in sidebar
The display or not of milestones in sidebar is included in the xml
export of the project.

At import, if the milestones are not displayed then the information is
recorded in database. If they are displayed, then no information is
recorded because it is the default behavior.

You can try by exporting a project with or with milestones in sidebar,
and importing it back. New project should have same settings. Remember
that you must have feature flags activated.

Part of story #35104: Milestones below Backlog

Change-Id: I36d695da2db6850c7801c1ef8f661b1eea2a02fc