chore(svn/el9): Move classes to SVNCore
This refactoring is needed to be able to introduce another refactoring
about the reading & writing of .SVNAccessFile with per repository rules.

In the future, the "default block" (Members and /) will be different
from one repository to another. However, as of today, the code that
reads and writes the default block is Project centric. We need to have
the Repository information in SVNCore in order to introduce the needed
refactoring.

Part of: request #31541 svn 1.10+ no longer supports duplicate sections

Change-Id: Ib61eaa5f0bbab64677f5c1e16d4e5e184ed5ea3e