feat: always display config in administration
In early development stage, checkbox about display of milestones in
sidebar was only displayed when checked. This was to hide the checkbox
from users because the feature is still under development.

Now that we have a feature flag dedicated to development mode, we can
always display the checkbox, checked or not, so that we can continue
development on this feature.

To test, have a project A with milestones in sidebar (1 or no entry in
plugin_agiledashboard_milestones_in_sidebar_config table) and a project
B without milestones in sidebar (0 in aforementioned table). If feature
flag is off, then you cannot see the checkbox in backlog administration
of the projects. If feature flag is on, then in project A the checkbox
is displayed checked, and in project B the checkbox is displayed
unchecked.

To activate or deactivate the feature flag:
tuleap config-set feature_flag_allow_milestones_in_sidebar_dev_mode 1
tuleap config-set feature_flag_allow_milestones_in_sidebar_dev_mode 0

Part of story #35104: Milestones below Backlog

Change-Id: I30e0a3c0e43b5e1f7693b6e4512d896f3ef1d3af