Use NodeJS 20.x with Tuleap Realtime
To test, build and deploy the 2 packages (`tuleap-node` and
`tuleap-realtime`) inside your dev env.

The code of Tuleap Realtime is now bundled using esbuild. This make
things easier to bring only the necessary code into the production
environment.

No functional changes.

We are forced to use an "unofficial" NodeJS build (which is not really
different than what we are doing right now) for CentOS 7. The glibc used
by CentOS 7 is too old to run the "standard" build binary.

Part of request #35159: Use NodeJS 20.x LTS for Tuleap node apps

Change-Id: I766bca4ff347b53745517454fd2db36c471dfbb3