feat: Inform admin about needed semantics
Administrators should know that in order to be able to see a milestone
in the sidebar, the semantics must be defined: timeframe, title, and
status.

It has been decided that only a simple message is displayed, no need to
block the update of the configuration if a semantic is missing.

Part of story #35104: Milestones below Backlog

Change-Id: I4239da2ae8836e74b932108c3032b044334ebda2