fix(svn/el9): Bring back historical permission scheme
In 1182b1ef816e9aa2b9d33ffa91be270196ba749d a change of behaviour was
introduced that made repository once world readable completly closed.

With this commit, we bring back the original behaviour because we don't
know the extend of the impact of this change (the code is running this
way since 10+ years) and there is no way to revert/adapt at repository
level because it's not possible to change [/] permissions yet.

Part of: request #35454 SVN repositories in public project are no longer public

Change-Id: Iff3546e3882e8b16e7a4ac692d00199078cc6847