feat: Add project dashboards in recent items
In order to be able to navigate easily to project dashboards, they
should be added in SwitchTo modal » Recent items.

Visits to a project dashboards are removed when the user clear its
history (API REST), when the dashboard is removed, or after daily
cleanup.

No functional changes in project dashboards CRUD (dao had to be
converted to EasyDB).

Closes request #35468: Add project dashboards in recent items

Change-Id: Ia92954fcb5e4c5db85ada1fabe788c55a549fdb0