chore: Drop Mockery usage in Request tests
This is part of request #21775 Favor PHPUnit mock system over Mockery - part 2

Last class is not modified because it involves at least one unit test that
mocks both existing method and one (myHandler) to create on the fly.

And the combination of ```addMethods``` and ```onlyMethods``` does not
seem to work in that case.

Change-Id: I348589564ca8ac53dc26e422e29986312855ea7c