feat: Remove dev mode feature flag
It's time to open the feature for early birds.

We have to truncate plugin_agiledashboard_milestones_in_sidebar_config
table because we need to reset the in sidebar config for projects
updating their scrum settings (with dev mode off, the project was marked
as without milestones in sidebar).

Part of story #35104: Milestones below Backlog

Change-Id: I7afb9186e76d05934a8ee326c9e463663c2f0569