Use Node 20.x to run Mathoid
To test build and redeploy the Mathoid package in your dev env.
You will also need to either update the `/usr/share/mediawiki-tuleap-flavor/current-lts/LocalSettings.Tuleap.php`
or to rebuild and redeploy the mediawiki-tuleap-flavor package.

Note that a patch is applied to the dependencies of Mathoid, their
`package.json` is incorrect as it indicates some packages that are
production dependencies as dev dependencies.

Part of request #35159: Use NodeJS 20.x LTS for Tuleap node apps

Change-Id: I4414d6223b915aea956f0a91a125447d32edefcd