fix: Check tracker hierarchy
Part of story #35104 Milestones below Backlog

Sometimes, two artifacts have a link '_is_child' but the trackers are
not parent/child in tracker_hierarchy. In this case we should not
display the child artifact.

*Testing:*

Create a release and a sprint. Both should be displayed in sidebar. If
you remove sprints from hierarchy of release, the sprint should be no
longer displayed.

Change-Id: Ia041360d8adb15902c7dd592310414ea866e7b73