fix: Don't get submilestones if user cannot see parent
Part of story #35104 Milestones below Backlog

If a user cannot see a parent artifact but can see its child, we should
not get the childs.

*Testing:*

Set permissions of Releases to only submitters and permissions of
Sprints to everyone. With a second user (not site-admin or
project-admin) you should see only the releases you have created with
their sprints.

Change-Id: Iaa501afd447ebd433fa4a02724d82f564e9b065a