refactor: Remove a fake empty time to display the table when it's empty
Part of request #35089 Convert personal-timetracking-widget to Vue3 + TS

Motivation:
When deleting the last time in the modal, an empty time (minute = null) was created to keep the modal display correct with the right artifact and project (because artifact and project were get with the first current time).
In addition to an architectural problem, this forces us to make dates and times nullable when migrating to TypeScript.

- Remove current_artifact and project_artifact getters
- Pass the objects to the component's children

No functional change expected, we can still add times after delete them all

Change-Id: I4f275d6635f7f500e515c7e322a4e995295c0806