chore: Remove Scrum mono milestone
epic #9422 Scrum v2 was drafted in 2016. A first MVP allowing a
"mono-milestone" mode was implemented and put behind a feature flag.
However, the epic was cancelled in 2021 and the experiment came to a
close. The MVP code and the feature flag (with its associated
conditions) should be removed. It will make it easier to maintain the
Backlog plugin (former AgileDashboard). Fewer conditions -> easier code.

The whole agiledashboard is not broken.

Closes request #34748: Remove Scrum Mono-milestone feature

Change-Id: I10991bbc515df99e87143f8cdbd5e3766545692f