fix: display both promoted tracker and kanban
In the sidebar, a tracker can't be promoted below "Tracker" if promoted
as a kanban. It should be interesting not to have that limitation.
Effectively, you may need to have both in your sidebar, not necessarily
for the same persona.

To test: have a promoted tracker, and create a promoted kanban based on
this tracker. Both tracker and kanban should appear in the sidebar.

Closes request #35491: [Sidebar] A tracker can't be promoted below "Tracker" if promoted as a kanban

Change-Id: Ic4056a65cb8a5c25aa73a04f434e2a707de77a1e