feat: display an empty state when there are no pull-requests
part of story #33006 Create filters on my PR

When there aren't any open, nor closed pull-requests in the repository,
we need to display an empty state.

In the mockups, this empty state contains the "create-pullrequest-button",
which means we cannot display it from the pull-request-homepage app,
because it is already a vue app, with its own mount-point.

How to test:
- Create a repository
- Clone and add a readme.md to it
- Checkout a new branch
- Make some changes in the readme file
- Push on a remote branch
- Go to the dashboad view of your repository
- In the url, add a parameter `tab=homepage`

--> There is an empty state saying "no pull-requests yet".
--> It contains the "create-pullrequest-button"

- Create a new PR from the button in the empty state, from the
  previously pushed branch into master

--> You are redirected to the overview page if the creation is a
success.
--> The empty state is no longer displayed if you go back to the
dashboard view (using `tab=homepage`).

- Fork this repository
--> In the fork, the dashboard view shows the empty state.
--> You can create a PR in the fork, or in its parent repository

Change-Id: I400edb88ebeb9eca6f9c7f6c1055e614dcd0460c