feat: display pull-requests reviewers
part of story #33006 Create filters on my PR

Note:

There aren't any indications in the mockup about the maximum number of
reviewers avatars to display. In order to have simpler css rules, I
chose to display them all.

We might need to ask UXD about it, because there aren't any wrapping
rule in this component, so the display can be really messed up if there
are too much reviewers.

How to test:

In a repository with at least two pull-requests:
- Assign some labels and reviewers to the first one
- Assign only reviewers to the second one

--> The reviewers avatars are align vertically
--> The reviewers avatars overlaps themselves

Change-Id: Ie0108f6d7e873345411ea64fe668fd04f6a09d4b